Refactor gridId type from uint128 to uint48 in WithdrawProfit events and related functions for improved efficiency

+++ src/facets/CancelFacet.sol
@@ -23,7 +23,7 @@ contract CancelFacet is IOrderEvents {
     using GridOrder for GridOrder.GridState;
     using SafeTransferLib for ERC20;
 
-    event WithdrawProfit(uint128 gridId, Currency quote, address to, uint256 amt);
+    event WithdrawProfit(uint48 gridId, Currency quote, address to, uint256 amt);
 
     error ETHTransferFailed();
     error NotWETH();
