+++ src/strategy/Geometry.sol
@@ -52,7 +52,7 @@ contract Geometry is IGridStrategy {
         GRID_EX = _gridEx;
     }
 
-    function gridIdKey(bool isAsk, uint128 gridId) internal pure returns (uint256) {
+    function gridIdKey(bool isAsk, uint48 gridId) internal pure returns (uint256) {
         if (isAsk) {
             return (uint256(ProtocolConstants.ASK_ORDER_FLAG) << 128) | uint256(gridId);
         }
